create Routes, routes package and refactor into Router
