--- src/main/java/com/github/martingaston/application/http/Handler.java
@@ -1,5 +1,0 @@
-package com.github.martingaston.application.http;
-
-interface Handler {
-    Response handle(Request request);
-}
