+++ src/test/java/com/github/martingaston/application/routes/RoutesTest.java
@@ -1,5 +1,7 @@
-package com.github.martingaston.application.http;
+package com.github.martingaston.application.routes;
 
+import com.github.martingaston.application.http.URI;
+import com.github.martingaston.application.http.Verbs;
 import org.junit.jupiter.api.*;
 
 import static org.assertj.core.api.Assertions.*;
